ch(Change user id from uuid to serial): Commit fixes the bug where the user id of type UUID can not be added to the database. [Fix #159917632]
